fps fixed together with xant

+++ src/include/fps.h
@@ -22,6 +22,7 @@
 #define __FPS_H__
 
 #include <pthread.h>
+#include <inttypes.h>
 
 class FPS {
  public:
@@ -33,7 +34,7 @@ class FPS {
   int get();
   int set(int rate);
   void calc();
-  void timeout(float delta);
+  void timeout(uint32_t delta);
   void delay();
 
   float fps, fps_old;
